+++ docs/release-notes.md
@@ -14,7 +14,7 @@ of `zizmor`.
 * zizmor now has **experimental** support for auditing pre-commit inputs,
   meaning both pre-commit configuration and hook definitions (#2209)
 
-* **New audit**: [insecure-origin] detects usages of insecure (i.e. plaintext)
+* **New audit**: [insecure-url-scheme] detects usages of insecure (i.e. plaintext)
   protocols when making network requests. The initial version of this audit
   is limited to pre-commit inputs only (#2228)
 
@@ -1995,6 +1995,7 @@ This is one of `zizmor`'s bigger recent releases! Key enhancements include:
 [typosquat-uses]: ./audits.md#typosquat-uses
 [unsound-ternary]: ./audits.md#unsound-ternary
 [adhoc-packages]: ./audits.md#adhoc-packages
+[insecure-url-scheme]: ./audits.md#insecure-url-scheme
 
 [exit code]: ./usage.md#exit-codes